+++ primefaces/src/main/frontend/packages/components/src/sticky/sticky.widget.js
@@ -21,6 +21,8 @@
  * @extends {PrimeFaces.widget.BaseWidgetCfg} cfg
  * 
  * @prop {number} cfg.margin Margin to the top of the page during fixed scrolling.
+ * @prop {string} cfg.stickyTopAt Selector for elements fixed at the top of the page whose height should be considered
+ * when positioning the sticky element.
  * @prop {string} cfg.target The client ID of the component to be made sticky.
  */
 PrimeFaces.widget.Sticky = class Sticky extends PrimeFaces.widget.BaseWidget {
@@ -34,11 +36,14 @@ PrimeFaces.widget.Sticky = class Sticky extends PrimeFaces.widget.BaseWidget {
         super.init(cfg);
         this.target = $(PrimeFaces.escapeClientId(this.cfg.target));
         this.cfg.margin = this.cfg.margin || 0;
+        this.cfg.stickyTopAt = this.cfg.stickyTopAt || null;
 
         this.initialState = {
             top: this.target.offset().top,
             height: this.target.height()
         };
+
+        this.calculateDynamicMargin();
 
         this.bindEvents();
     }
@@ -71,7 +76,22 @@ PrimeFaces.widget.Sticky = class Sticky extends PrimeFaces.widget.BaseWidget {
                 $this.restore();
         });
 
-        PrimeFaces.utils.registerResizeHandler(this, 'resize.' + this.id + '_align', null, function() {
+        PrimeFaces.utils.registerResizeHandler(this, 'resize.' + this.id + '_align', null, function () {
+            var marginChanged = $this.calculateDynamicMargin();
+            $this.initialState.top = ($this.ghost && $this.ghost.length && $this.ghost.offset()
+                    && $this.ghost.offset().top > 0)
+                    ? $this.ghost.offset().top
+                    : ($this.target.offset() ? $this.target.offset().top : $this.initialState.top);
+            if ($(window).scrollTop() > $this.initialState.top - $this.cfg.margin) {
+                if (marginChanged && $this.fixed) {
+                    $this.ghost.remove();
+                    $this.fix(true);
+                } else {
+                    $this.fix();
+                }
+            } else {
+                $this.restore();
+            }
             if ($this.fixed) {
                 $this.target.width($this.ghost.outerWidth() - ($this.target.outerWidth() - $this.target.width()));
             }
@@ -115,6 +135,20 @@ PrimeFaces.widget.Sticky = class Sticky extends PrimeFaces.widget.BaseWidget {
             this.ghost.remove();
             this.fixed = false;
         }
+    }
+
+    /**
+     * @return {boolean} true if the margin has been changed, false otherwise
+     */
+    calculateDynamicMargin() {
+        var fixedElementsOnTop = this.cfg.stickyTopAt ? $(this.cfg.stickyTopAt) : null;
+        var previousMargin = this.cfg.margin;
+        if (fixedElementsOnTop && fixedElementsOnTop.length) {
+            this.cfg.margin = fixedElementsOnTop.toArray().reduce(function(height, elem) {
+                return height + $(elem).outerHeight();
+            }, 0);
+        }
+        return previousMargin !== this.cfg.margin;
     }
 
 }